feat: add working prototype for user page
